+++ tests/test_utils.py
@@ -61,26 +61,6 @@ class TestSplitCodeAndComment:
         assert split == stripped_code_and_comment
 
 
-class TestSurrounds:
-    @staticmethod
-    def test_should_return_true_for_error_within_region() -> None:
-        region = UnsilenceableRegion((1, 0), (2, 0))
-        error = MypyError("", 0, 1, "", "")
-        assert region.surrounds(error)
-
-    @staticmethod
-    def test_should_return_true_for_error_within_infinite_region() -> None:
-        region = UnsilenceableRegion((1, 0), (2, -1))
-        error = MypyError("", 0, 1, "", "")
-        assert region.surrounds(error)
-
-    @staticmethod
-    def test_should_return_false_for_error_outside_region() -> None:
-        region = UnsilenceableRegion((1, 0), (2, 0))
-        error = MypyError("", 3, 0, "", "")
-        assert not region.surrounds(error)
-
-
 class TestFindUnsilenceableRegions:
     @staticmethod
     def test_should_return_explicitly_continued_lines() -> None:
@@ -94,7 +74,7 @@ class TestFindUnsilenceableRegions:
         )
         stream = io.StringIO(code)
         regions = find_unsilenceable_regions(stream)
-        expected = UnsilenceableRegion((1, 0), (1, -1))
+        expected = UnsilenceableRegion((1, 0), (1, 8))
         assert expected in regions
 
     @staticmethod
@@ -125,7 +105,7 @@ class TestFindSafeEndLine:
         None
     ):
         error = MypyError("", 0, 1, "", "")
-        region = UnsilenceableRegion((1, 0), (1, -1))
+        region = UnsilenceableRegion((1, 0), (1, 1))
         end_line = find_safe_end_line(error, [region])
         assert end_line == -1